Move panel test to frontend
