+++ tests/components/test_websocket_api.py
@@ -7,29 +7,12 @@ from async_timeout import timeout
 import pytest
 
 from homeassistant.core import callback
-from homeassistant.components import websocket_api as wapi, frontend
+from homeassistant.components import websocket_api as wapi
 from homeassistant.setup import async_setup_component
 
 from tests.common import mock_coro
 
 API_PASSWORD = 'test1234'
-
-
-@pytest.fixture
-def websocket_client(loop, hass, aiohttp_client):
-    """Websocket client fixture connected to websocket server."""
-    assert loop.run_until_complete(
-        async_setup_component(hass, 'websocket_api'))
-
-    client = loop.run_until_complete(aiohttp_client(hass.http.app))
-    ws = loop.run_until_complete(client.ws_connect(wapi.URL))
-    auth_ok = loop.run_until_complete(ws.receive_json())
-    assert auth_ok['type'] == wapi.TYPE_AUTH_OK
-
-    yield ws
-
-    if not ws.closed:
-        loop.run_until_complete(ws.close())
 
 
 @pytest.fixture
@@ -287,31 +270,6 @@ def test_get_config(hass, websocket_client):
             set(msg['result']['whitelist_external_dirs'])
 
     assert msg['result'] == hass.config.as_dict()
-
-
-@asyncio.coroutine
-def test_get_panels(hass, websocket_client):
-    """Test get_panels command."""
-    yield from hass.components.frontend.async_register_built_in_panel(
-        'map', 'Map', 'mdi:account-location')
-    hass.data[frontend.DATA_JS_VERSION] = 'es5'
-    yield from websocket_client.send_json({
-        'id': 5,
-        'type': 'get_panels',
-    })
-
-    msg = yield from websocket_client.receive_json()
-    assert msg['id'] == 5
-    assert msg['type'] == wapi.TYPE_RESULT
-    assert msg['success']
-    assert msg['result'] == {'map': {
-        'component_name': 'map',
-        'url_path': 'map',
-        'config': None,
-        'url': None,
-        'icon': 'mdi:account-location',
-        'title': 'Map',
-    }}
 
 
 @asyncio.coroutine
